It _considers_ them
I added constraints. Now I just have to satisfy them.

I did a bunch of the mathematical legwork for constraints, the only thing left is to actually solve for and apply the restorative forces. And fix it, because there's no way it's not hopelessly broken now. I just wanted to commit now while it runs.