Hinges!
There are hinges now! Turns out all I needed to avoid the non-singular matrix issue was to put a "p" in front of the "inv" in the constraint solving code.

Of course, the pseudo-inverse of a 12x12 is _way_ slower than the true inverse of a 10x10, so I'm going to keep pursuing parallel joints, anyway. Also, the physics seem off, so I'm going to check on my response matrices.