It's struggling, but it's doing it.
There are hinges. You can tell these equations aren't easy to solve, and the drift definitely gets very noticeable at times unless you turn the recovery time way down. But it clearly _works_! This is exciting. Now all that's left is to encapsulate the ball and parallel joint in a convenient little class.