Yup, that's rotation.
I've been validating the rotation code. There turned out to be a quaternion-magnitude-drift-induced instability that I extinguished, which also involved me finding out that pyquaternion has been normalizing my quaternions without me realizing it in parts of the code. Point is, the rotation works better now, and I'm pretty confidant in it.