*efficiency*
I realised the GMAT integration was super inefficient so I replaced the sweep with a binary search and shed a few seconds off a simulation of 10 minutes. I also verified that the tables are behaving as one would expect. For my last trick, I think I want the magnetic field to adjust according to the relative location of the Earth.