I have a theory.
My simulations suddenly got extremely slow when I strengthened the magnetorquers, and I think it may be that the rate at which the magnetorquers change direction approaches infinity when they get very stable. I'm making them go linear when the rate of rotation is very low to see if it speeds things up.